rename `AnyMap` → `TypeMap`
