+++ egui/src/any/type_map.rs
@@ -1,16 +1,6 @@
+use crate::any::element::{AnyMapElement, AnyMapTrait};
+use std::any::TypeId;
 use std::collections::HashMap;
-
-#[cfg(feature = "persistence")]
-use {
-    crate::any::serializable::element::{AnyMapElement, AnyMapTrait},
-    crate::any::serializable::type_id::TypeId,
-};
-
-#[cfg(not(feature = "persistence"))]
-use {
-    crate::any::element::{AnyMapElement, AnyMapTrait},
-    std::any::TypeId,
-};
 
 /// Maps types to a single instance of that type.
 ///
@@ -18,14 +8,13 @@ use {
 /// Similar to [the `typemap` crate](https://docs.rs/typemap/0.3.3/typemap/) but allows serialization
 /// (if compiled with the `persistence` feature).
 #[derive(Clone, Debug, Default)]
-#[cfg_attr(feature = "persistence", derive(serde::Deserialize, serde::Serialize))]
-pub struct AnyMap(HashMap<TypeId, AnyMapElement>);
+pub struct TypeMap(HashMap<TypeId, AnyMapElement>);
 
 // ----------------------------------------------------------------------------
 
-impl AnyMap {
-    pub fn get<T: AnyMapTrait>(&mut self) -> Option<&T> {
-        self.get_mut().map(|x| &*x)
+impl TypeMap {
+    pub fn get<T: AnyMapTrait>(&self) -> Option<&T> {
+        self.0.get(&TypeId::of::<T>())?.get()
     }
 
     pub fn get_mut<T: AnyMapTrait>(&mut self) -> Option<&mut T> {
@@ -33,7 +22,7 @@ impl AnyMap {
     }
 }
 
-impl AnyMap {
+impl TypeMap {
     pub fn get_or_insert_with<T: AnyMapTrait>(&mut self, or_insert_with: impl FnOnce() -> T) -> &T {
         &*self.get_mut_or_insert_with(or_insert_with)
     }
@@ -61,14 +50,14 @@ impl AnyMap {
     }
 }
 
-impl AnyMap {
+impl TypeMap {
     pub fn insert<T: AnyMapTrait>(&mut self, element: T) {
         self.0
             .insert(TypeId::of::<T>(), AnyMapElement::new(element));
     }
 }
 
-impl AnyMap {
+impl TypeMap {
     pub fn reset<T: AnyMapTrait>(&mut self) {
         self.0.remove(&TypeId::of::<T>());
     }
