+++ egui/src/memory.rs
@@ -19,11 +19,11 @@ pub struct Memory {
     pub options: Options,
 
     /// This map stores current states for widgets that don't require `Id`. This will be saved between different program runs if you use the `persistence` feature.
-    pub data: any::serializable::AnyMap,
+    pub data: any::serializable::TypeMap,
 
     /// Same as `data`, but this data will not be saved between runs.
     #[cfg_attr(feature = "persistence", serde(skip))]
-    pub data_temp: any::AnyMap,
+    pub data_temp: any::TypeMap,
 
     /// This map stores current states for all widgets with custom `Id`s. This will be saved between different program runs if you use the `persistence` feature.
     pub id_data: any::serializable::AnyMapId,
